requirements.txt
Flask==2.3.3
pandas==2.1.1
plotly==5.17.0
prophet==1.1.4
numpy==1.24.3
Werkzeug==2.3.7
openpyxl==3.1.2
python-dotenv==1.0.0 